共同開発：最終調整

+++ resources/views/welcome.blade.php
@@ -17,22 +17,22 @@
                         <a class="no-text-decoration" href="{{ route('show', $post->user->id) }}">
                             <i class="fas fa-user-circle fa-2x mr-1"></i>
                         </a>
-                        <a class="black_color" title="" href="{{ route('show', $post->user->id) }}">
+                        <a class="black-color" title="" href="{{ route('show', $post->user->id) }}">
                             <strong>
                                 {{ $post->user->name }}
                             </strong>
                         </a>
                     </div>
                     <div class="card-body">
-                        <div class="post_edit text-right" style="display: flex; justify-content: flex-end;">
-                            <a class="btn btn-primary btn-sm" href="{{ route('edit', $post->id) }}"><i
-                                    class="far fa-edit"></i>編集</a>
-                            {{-- @if (Auth::id() == $post->user_id) --}}
-                            {!! Form::open(['route' => ['delete', $post->id], 'method' => 'delete']) !!}
-                            {!! Form::button('<i class="far fa-trash-alt" style="margin-left: 2px;"></i>削除', ['class' => 'btn btn-danger btn-sm', 'type' => 'submit']) !!}
-                            {!! Form::close() !!}
-                            {{-- @endif --}}
-                        </div>
+                        @if (Auth::id() == $post->user_id)
+                            <div class="post_edit text-right" style="display: flex; justify-content: flex-end;">
+                                <a class="btn btn-primary btn-sm" href="{{ route('edit', $post->id) }}"><i
+                                        class="far fa-edit"></i>編集</a>
+                                {!! Form::open(['route' => ['delete', $post->id], 'method' => 'delete']) !!}
+                                {!! Form::button('<i class="far fa-trash-alt" style="margin-left: 2px;"></i>削除', ['class' => 'btn btn-danger btn-sm', 'type' => 'submit']) !!}
+                                {!! Form::close() !!}
+                            </div>
+                        @endif
                         <h3 class="h5 title">
                             {{ $post->title }}
                         </h3>
